% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muac_missing_data.R
\name{check_missing_data}
\alias{check_missing_data}
\title{Check missing anthropometry data}
\usage{
check_missing_data(df)
}
\arguments{
\item{df}{A data.frame with information on age, sex, oedema status, and
MUAC of each child that has been processed using \code{process_muac_data()}}
}
\value{
A tibble summarising number and percent missing data for age, sex,
oedema status, and MUAC for the given \code{df}
}
\description{
Check missing anthropometry data
}
\examples{
check_missing_data(muac_data)

}
