% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muac_utils.R
\name{classify_age_ratio}
\alias{classify_age_ratio}
\alias{classify_sex_ratio}
\alias{classify_sd}
\alias{classify_quality}
\alias{classify_acute_malnutrition}
\title{Utility functions that support the main functions in the package}
\usage{
classify_age_ratio(p)

classify_sex_ratio(p)

classify_sd(std_dev)

classify_quality(age_ratio_class, sex_ratio_class, std_dev_class, dps_class)

classify_acute_malnutrition(
  muac,
  muac_units = c("mm", "cm"),
  oedema,
  oedema_recode = NULL
)
}
\arguments{
\item{p}{Numeric value for p-value of a statistical test used in the
various checks applied.}

\item{std_dev}{Numeric value for standard deviation (SD) of a measurement
usually MUAC.}

\item{age_ratio_class}{A character value or vector for classification based
on the result of the age ratio test.}

\item{sex_ratio_class}{A character value or vector for classification based
on the sex ratio test.}

\item{std_dev_class}{A character value for vector for classification based
on standard deviation.}

\item{dps_class}{A character value for vector for classification based on
the digit preference score (DPS)}

\item{muac}{A numeric value or vector of numeric values for MUAC measurement
of child. The expected values for MUAC are in millimetres. If units are
different, use \code{muac_units} to specify which units are used.}

\item{muac_units}{A character value for units used for MUAC measurement.
Currently accepts either "mm" for millimetres (default) or "cm" for
centimetres.}

\item{oedema}{A value or a vector of values for oedema status of child. The
expected values for \code{oedema} is 1 = for presence of oedema and 2 for no
oedema. If data values are different, use \code{oedema_recode} to map out the
values to what is required.}

\item{oedema_recode}{A vector of values with length of 2 with the first
element for the value signifiying presence of oedema and second element for
the value signifying no oedema in the dataset. For example, if "y" is the
value for presence of oedema and "n" is the value for no oedema, then
specify \verb{c("y", "n)}. If set to NULL (default), then the values c(1, 0)
are used.}
}
\value{
A single value or a vector of values providing a classification
}
\description{
Utility functions that support the main functions in the package
}
\examples{
age_ratio_p <- nipnTK::ageRatioTest(as.integer(!is.na(muac_data$age)))$p
classify_age_ratio(age_ratio_p)

}
