% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muac_checks.R
\name{ipc_muac_check}
\alias{ipc_muac_check}
\title{Perform MUAC check based on IPC and CDC recommendations}
\usage{
ipc_muac_check(
  df,
  age = "age",
  sex = "sex",
  sex_recode = NULL,
  muac = "muac",
  muac_units = c("mm", "cm"),
  oedema = "oedema",
  oedema_recode = NULL,
  .summary = TRUE
)
}
\arguments{
\item{df}{A data.frame with information on age, sex, oedema status, and
MUAC of each child}

\item{age}{A character value for name of variable in `df` for age of
child. The age of child should be in months.}

\item{sex}{A character value for name of variale in `df` for sex of child.
The expected values for `sex` is 1 = males; 2 = females. If data values
are different, use `sex_recode` to map out the values to what is
required.}

\item{sex_recode}{A vector of values with length of 2 with the first
element for the value signifiying males and second element for the value
signifying females in the dataset. For example, if "m" is the value for
males and "f" is the value for females, then specify `c("m", "f)`. If
set to NULL (default), then the values c(1, 2) are used.}

\item{muac}{A character value for name of variable in `df` for MUAC
measurement of child. The expected values for MUAC are in millimetres.
If units are different, use `muac_units` to specify which units are used.}

\item{muac_units}{A character value for units used for MUAC measurement.
Currently accepts either "mm" for millimetres (default) or "cm" for
centimetres.}

\item{oedema}{A character value for name of variable in `df` for oedema
status of child. The expected values for `oedema` is 1 = for presence of
oedema and 2 for no oedema. If data values are different, use
`oedema_recode` to map out the values to what is required.}

\item{oedema_recode}{A vector of values with length of 2 with the first
element for the value signifiying presence of oedema and second element for
the value signifying no oedema in the dataset. For example, if "y" is the
value for presence of oedema and "n" is the value for no oedema, then
specify `c("y", "n)`. If set to NULL (default), then the values c(1, 0)
are used.}

\item{.summary}{Logical. Should output be a summary of all the checks
peformed on the MUAC dataset? If TRUE (default), output will be a single
row data.frame with each column for each metric used to check MUAC
dataset. If FALSE, a data.frame with same number of rows as `df` and
columns for each metric used to check MUAC dataset is added to `df`.
Setting `.summary` to FALSE is usually only used for when the output
structure is required for further analysis (i.e., calculation of
prevalence).}
}
\value{
A data.frame with a single row with each column for each metric used
  to check MUAC dataset if `.summary` is TRUE. If `.summary` is FALSE, a
  data.frame with same number of rows as `df` and columns for each metric
  used to check MUAC dataset is added to `df`.
}
\description{
Perform MUAC check based on IPC and CDC recommendations
}
\examples{
ipc_muac_check(df = muac_data, oedema_recode = c(1, 2), muac_units = "cm")

}
