% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipctools.R
\docType{package}
\name{ipctools}
\alias{ipctools}
\alias{ipctools-package}
\title{Utilities to Support Integrated Food Security Phase Classification (IPC) Data
Analysis and Visualisation}
\description{
The Integrated Food Security Phase Classification (IPC) is a widely used tool
for classifying and analyzing the severity and magnitude of food insecurity
and malnutrition situations in various countries and regions around the
world. It provides a common understanding of the food security situation and
enables decision-makers to take appropriate actions to mitigate and respond
to food crises. This package provides functions and utilities that support
IPC-related data analysis and visualisation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nutriverse.io/ipctools/}
  \item \url{https://github.com/nutriverse/ipctools}
  \item Report bugs at \url{https://github.com/nutriverse/ipctools/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Tomas Zaba [copyright holder]
}

}
\keyword{internal}
