% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{muac_data}
\alias{muac_data}
\title{Nutrition survey data with MUAC and oedema measurements and location
information anonymised.}
\format{
A data frame with 6 columns and 435 rows:

| **Variable** | **Description** |
| :--- | :--- |
| *state_name* | Name of state |
| *district_name* | Name of district |
| *age* | Age of child in months |
| *sex* | Sex of child. 1 = male; 2 = female |
| *muac* | Mid-upper arm circumference (MUAC) measurement in centimetres |
| *oedema* | Presence or absence of oedema. 1 = oedema; 2 = no oedema |
}
\usage{
muac_data
}
\description{
Nutrition survey data with MUAC and oedema measurements and location
information anonymised.
}
\examples{
muac_data


}
\keyword{datasets}
