% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_muac.R
\name{process_muac_data}
\alias{process_muac_data}
\title{Process MUAC data}
\usage{
process_muac_data(
  df,
  age = "age",
  sex = "sex",
  sex_recode = NULL,
  muac = "muac",
  muac_units = c("mm", "cm"),
  oedema = "oedema",
  oedema_recode = NULL
)
}
\arguments{
\item{df}{A data.frame with information on age, sex, oedema status, and
MUAC of each child}

\item{age}{A character value for name of variable in \code{df} for age of
child. The age of child should be in months.}

\item{sex}{A character value for name of variable in \code{df} for sex of child.
The expected values for \code{sex} is 1 = males; 2 = females. If data values
are different, use \code{sex_recode} to map out the values to what is
required.}

\item{sex_recode}{A vector of values with length of 2 with the first
element for the value signifying males and second element for the value
signifying females in the dataset. For example, if "m" is the value for
males and "f" is the value for females, then specify \verb{c("m", "f)}. If
set to NULL (default), then the values c(1, 2) are used.}

\item{muac}{A character value for name of variable in \code{df} for MUAC
measurement of child. The expected values for MUAC are in millimetres.
If units are different, use \code{muac_units} to specify which units are used.}

\item{muac_units}{A character value for units used for MUAC measurement.
Currently accepts either "mm" for millimetres (default) or "cm" for
centimetres.}

\item{oedema}{A character value for name of variable in \code{df} for oedema
status of child. The expected values for \code{oedema} is 1 = for presence of
oedema and 2 for no oedema. If data values are different, use
\code{oedema_recode} to map out the values to what is required. If dataset
does not have oedema values, set this to NULL.}

\item{oedema_recode}{A vector of values with length of 2 with the first
element for the value signifying presence of oedema and second element for
the value signifying no oedema in the dataset. For example, if "y" is the
value for presence of oedema and "n" is the value for no oedema, then
specify \verb{c("y", "n)}. If set to NULL (default), then the values c(1, 0)
are used.}
}
\value{
An appropriately structured data.frame that can be passed on to
other functions
}
\description{
Process MUAC data
}
\examples{
process_muac_data(muac_data)

}
