% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muac_prevalence.R
\name{calculate_unweighted_prevalence}
\alias{calculate_unweighted_prevalence}
\alias{calculate_weighted_prevalence}
\alias{calculate_prevalence}
\title{Calculate wasting prevalence by MUAC}
\usage{
calculate_unweighted_prevalence(
  muac,
  muac_units = c("mm", "cm"),
  oedema,
  oedema_recode = NULL,
  status = c("sam", "mam")
)

calculate_weighted_prevalence(
  age,
  sex,
  sex_recode = NULL,
  muac,
  muac_units = c("mm", "cm"),
  oedema,
  oedema_recode = NULL,
  status = c("sam", "mam")
)

calculate_prevalence(
  age,
  sex,
  sex_recode = NULL,
  muac,
  muac_units = c("mm", "cm"),
  oedema,
  oedema_recode = NULL,
  quality_class,
  status = c("sam", "mam")
)
}
\arguments{
\item{muac}{A numeric value or vector of numeric values for MUAC measurement
of child. The expected values for MUAC are in millimetres. If units are
different, use `muac_units` to specify which units are used.}

\item{muac_units}{A character value for units used for MUAC measurement.
Currently accepts either "mm" for millimetres (default) or "cm" for
centimetres.}

\item{oedema}{A value or a vector of values for oedema status of child. The
expected values for `oedema` is 1 = for presence of oedema and 2 for no
oedema. If data values are different, use `oedema_recode` to map out the
values to what is required.}

\item{oedema_recode}{A vector of values with length of 2 with the first
element for the value signifiying presence of oedema and second element for
the value signifying no oedema in the dataset. For example, if "y" is the
value for presence of oedema and "n" is the value for no oedema, then
specify `c("y", "n)`. If set to NULL (default), then the values c(1, 0)
are used.}

\item{status}{Which wasting anthropometric indicator to report. A choice
between c("sam", "mam"). Default to "sam"}

\item{age}{A numeric or integer value or vector of values for age of child.
The age of child should be in months.}

\item{sex}{A value or a vector of values for sex of child. The expected
values for `sex` is 1 = males; 2 = females. If data values are different,
use `sex_recode` to map out the values to what is required.}

\item{sex_recode}{A vector of values with length of 2 with the first
element for the value signifiying males and second element for the value
signifying females in the dataset. For example, if "m" is the value for
males and "f" is the value for females, then specify `c("m", "f)`. If
set to NULL (default), then the values c(1, 2) are used.}

\item{quality_class}{A vector or character values indicating the
classification of the quality of the MUAC dataset. This is usually
created from applying the `ipc_muac_check()` function.}
}
\value{
A single value, a vector of values, or a table providing a prevalence
}
\description{
Calculate wasting prevalence by MUAC
}
\examples{
calculate_unweighted_prevalence(
  muac = muac_data$muac,
  oedema = muac_data$oedema,
  status = "sam"
)

df <- ipc_muac_check(
  muac_data, age = "age", sex = "sex",
  muac = "muac", muac_units = "cm",
  oedema = "oedema", oedema_recode = c(1, 2),
  .summary = FALSE
)

with(df,
  calculate_prevalence(
    age = age, sex = sex, muac = muac,
    oedema = oedema, quality_class = quality_class
  )
)

}
