% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{muac_data}
\alias{muac_data}
\title{Nutrition survey data with MUAC and oedema measurements and location
information anonymised.}
\format{
A data frame with 6 columns and 435 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{state_name} \tab Name of state \cr
   \emph{district_name} \tab Name of district \cr
   \emph{age} \tab Age of child in months \cr
   \emph{sex} \tab Sex of child. 1 = male; 2 = female \cr
   \emph{muac} \tab Mid-upper arm circumference (MUAC) measurement in centimetres \cr
   \emph{oedema} \tab Presence or absence of oedema. 1 = oedema; 2 = no oedema \cr
}
}
\usage{
muac_data
}
\description{
Nutrition survey data with MUAC and oedema measurements and location
information anonymised.
}
\examples{
muac_data


}
\keyword{datasets}
